Remove unnecessary stuff from walk replacement
The original walk was a Shortcut replacement for all uses of
clojure.walk/walk and had additional enhancements.
The only thing qslice needs is metadata preservation,
so this commit strips out unrelated enhancements,
chiefly the ArraySeq specialization and its tests.
There is already a qslice-core test that ensures metadata is preserved.

Also note that the metadata preservation bug is fixed in clojure 1.12;
qslice.walk is now only for compatibility with older clojure versions.